+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## Minor Changes
+- Implement write methods for Apugli's SerializableDataTypes. ([#63](https://github.com/MerchantPug/apugli/issues/63)
+
 ## Bugfixes
 - Fixed `sprinting` power type deactivating when an entity is touching water but is not underwater.